recipe/macos-before-10.13.patch: reorder some includes to fix macOS build
It seems that in macOS < 10.13, the type NSPasteboardType needs to be
manually defined, which is done in
`gdk/macos/gdkmacosclipboard-private.h`. But
`.../gdkmacospasteboard-private.h` is included from that header and
references the type before it is defined. So in this patch we change the
ordering a bit to fix things.